Moved helper functions to their own package
