+++ src/main/java/hsvi/bounds/SawtoothUpperBound.java
@@ -9,17 +9,21 @@ public class SawtoothUpperBound extends UpperBound {
 
     public SawtoothUpperBound(int dimension, List<UBPoint> initialUBPoints) {
         this(dimension);
-
+        initUBPoints(initialUBPoints);
     }
 
     @Override
     public void addPoint(UBPoint point, int a) {
-
+        points.add(point);
+        if (1 - (double)lastPrunedSize / size() >= pruningGrowthRatio) {
+            removeDominated();
+            lastPrunedSize = size();
+        }
     }
 
     @Override
     public void addPoint(double[] belief, double value, int a) {
-
+        addPoint(new UBPoint(belief, value, a));
     }
 
     @Override
@@ -29,6 +33,6 @@ public class SawtoothUpperBound extends UpperBound {
 
     @Override
     public void removeDominated() {
-
+        System.out.println("Removing dominated");
     }
 }
